+++ R/misc.R
@@ -60,3 +60,9 @@ na_splice <- function(new_data, synthetic_data, object) {
 
   res[, names(new_data)]
 }
+
+#https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode
+Mode <- function(x) {
+  ux <- unique(x)
+  ux[which.max(tabulate(match(x, ux)))]
+}