+++ NEWS.md
@@ -4,6 +4,7 @@
 * `step_tomek()` now ignore non-predictor variables when appropriate. (#51)
 * Update examples to no longer use `iris` or `okc` data sets.
 * All recipe steps now officially support empty selections to be more aligned with dplyr and other packages that use tidyselect (#55)
+* `step_smotenc()` have been added to implement SMOTENC which can handle categorical as well as numerical values. Thanks to @RobertGregg (#82)
 
 # themis 0.1.4